+++ operators/multiclusterobservability/controllers/multiclusterobservability/multiclusterobservability_controller.go
@@ -49,6 +49,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	mcov1beta2 "github.com/stolostron/multicluster-observability-operator/operators/multiclusterobservability/api/v1beta2"
+	analyticsctrl "github.com/stolostron/multicluster-observability-operator/operators/multiclusterobservability/controllers/multiclusterobservability/analytics"
 	placementctrl "github.com/stolostron/multicluster-observability-operator/operators/multiclusterobservability/controllers/placementrule"
 	certctrl "github.com/stolostron/multicluster-observability-operator/operators/multiclusterobservability/pkg/certificates"
 	"github.com/stolostron/multicluster-observability-operator/operators/multiclusterobservability/pkg/config"
@@ -398,8 +399,8 @@ func (r *MultiClusterObservabilityReconciler) Reconcile(ctx context.Context, req
 		}
 	}
 
-	// create analytics component
-	result, err = CreateAnalyticsComponent(r.Client, r.Scheme, instance, r.Manager)
+	// create rightsizing component
+	result, err = analyticsctrl.CreateRightSizingComponent(r.Client, r.Scheme, instance, r.Manager)
 	if result != nil {
 		return *result, err
 	}
@@ -496,7 +497,7 @@ func (r *MultiClusterObservabilityReconciler) SetupWithManager(mgr ctrl.Manager)
 
 	mcoPred := GetMCOPredicateFunc()
 	cmPred := GetConfigMapPredicateFunc()
-	cmNamespaceRSPred := getNamespaceRSConfigMapPredicateFunc(c)
+	cmNamespaceRSPred := analyticsctrl.GetNamespaceRSConfigMapPredicateFunc(c)
 	secretPred := GetAlertManagerSecretPredicateFunc()
 	namespacePred := GetNamespacePredicateFunc()
 	mcoaCRDPred := GetMCOACRDPredicateFunc()
